Change unit tests to compile at run-time.
This is done because the form types are determine during
macroexpansion, in order to have access to the lexical
environment. Without this re-running the test will not call FORM-TYPE
again.

Additionally some implementations, like ECL, only store the type
information in the environment when the test is compiled at
run-time. This is due to COMPILE being used when the test is compiled
at run-time, rather than EVAL.